Move data sources into lenskit-core (#675)
