+++ lenskit-core/src/main/java/org/grouplens/lenskit/data/source/CSVDataSourceBuilder.java
@@ -18,7 +18,7 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.grouplens.lenskit.eval.data;
+package org.grouplens.lenskit.data.source;
 
 import com.google.common.base.Preconditions;
 import org.apache.commons.lang3.builder.Builder;
