+++ lenskit-eval/src/main/java/org/grouplens/lenskit/eval/TrainModelTask.java
@@ -26,8 +26,8 @@ import org.apache.commons.lang3.time.StopWatch;
 import org.grouplens.lenskit.RecommenderBuildException;
 import org.grouplens.lenskit.core.LenskitConfiguration;
 import org.grouplens.lenskit.core.LenskitRecommender;
+import org.grouplens.lenskit.data.source.DataSource;
 import org.grouplens.lenskit.eval.algorithm.AlgorithmInstance;
-import org.grouplens.lenskit.eval.data.DataSource;
 import org.grouplens.lenskit.util.LogContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
