+++ lenskit-eval/src/main/java/org/grouplens/lenskit/eval/data/crossfold/CrossfoldTask.java
@@ -34,10 +34,10 @@ import org.grouplens.lenskit.data.dao.packed.BinaryFormatFlag;
 import org.grouplens.lenskit.data.event.Rating;
 import org.grouplens.lenskit.data.history.UserHistory;
 import org.grouplens.lenskit.data.pref.Preference;
+import org.grouplens.lenskit.data.source.CSVDataSourceBuilder;
+import org.grouplens.lenskit.data.source.DataSource;
 import org.grouplens.lenskit.eval.AbstractTask;
 import org.grouplens.lenskit.eval.TaskExecutionException;
-import org.grouplens.lenskit.eval.data.CSVDataSourceBuilder;
-import org.grouplens.lenskit.eval.data.DataSource;
 import org.grouplens.lenskit.eval.data.RatingWriter;
 import org.grouplens.lenskit.eval.data.RatingWriters;
 import org.grouplens.lenskit.eval.data.pack.PackedDataSourceBuilder;
