+++ lenskit-eval/src/main/resources/META-INF/lenskit-eval/builders.properties
@@ -1,1 +1,2 @@
 org.grouplens.lenskit.data.pref.PreferenceDomain=org.grouplens.lenskit.data.pref.PreferenceDomainBuilder
+org.grouplens.lenskit.data.source.DataSource=org.grouplens.lenskit.data.source.CSVDataSourceBuilder
