+++ lenskit-eval/src/main/resources/META-INF/lenskit-eval/methods/csvfile.properties
@@ -1,1 +1,1 @@
-builder=org.grouplens.lenskit.eval.data.CSVDataSourceBuilder
+builder=org.grouplens.lenskit.data.source.CSVDataSourceBuilder
