+++ lenskit-eval/src/test/groovy/org/grouplens/lenskit/eval/TrainModelTaskTest.groovy
@@ -25,7 +25,7 @@ import org.grouplens.lenskit.baseline.GlobalMeanRatingItemScorer
 import org.grouplens.lenskit.baseline.ItemMeanRatingItemScorer
 import org.grouplens.lenskit.data.dao.EventCollectionDAO
 import org.grouplens.lenskit.data.event.Ratings
-import org.grouplens.lenskit.eval.data.GenericDataSource
+import org.grouplens.lenskit.data.source.GenericDataSource
 import org.grouplens.lenskit.eval.script.ConfigTestBase
 import org.junit.Test
