+++ lenskit-eval/src/test/java/org/grouplens/lenskit/eval/data/traintest/GenericTTDataBuilderTest.java
@@ -22,7 +22,7 @@ package org.grouplens.lenskit.eval.data.traintest;
 
 import org.grouplens.lenskit.data.dao.EventCollectionDAO;
 import org.grouplens.lenskit.data.event.Rating;
-import org.grouplens.lenskit.eval.data.GenericDataSource;
+import org.grouplens.lenskit.data.source.GenericDataSource;
 import org.junit.Test;
 
 import java.util.Collections;
